+++ EntradaSalida/Perifericos/Displays7Seg/AgrupadorDigitos/INTERFAZ_EstrategiaPadding.h
@@ -1,7 +1,7 @@
 #ifndef ESTRATEGIA_PADDING
 #define ESTRATEGIA_PADDING
 
-#define DISPLAY7SEG_APAGADO 255
+#define DISPLAY7SEG_APAGADO ((byte) 255)
 
 #include <Array.h>
     template <size_t NumDigitos>
